fallback

+++ src/optimum/rbln/transformers/models/blip_2/modeling_blip_2.py
@@ -474,9 +474,7 @@ class RBLNBlip2ForConditionalGeneration(RBLNModel, RBLNDecoderOnlyGenerationMixi
             A list of strings of length batch_size * num_captions.
         """
         batch_size = pixel_values.shape[0]
-        
-        # FIXME(seinpark): is it available for type casting in each submodules?
-        pixel_values = pixel_values.to(self.vision_model.dtype) # ours
+
         image_embeds = self.vision_model(
             pixel_values,
             return_dict=True,
@@ -485,8 +483,6 @@ class RBLNBlip2ForConditionalGeneration(RBLNModel, RBLNDecoderOnlyGenerationMixi
         image_attention_mask = torch.ones(image_embeds.size()[:-1], dtype=torch.long, device=image_embeds.device)
 
         query_tokens = self.query_tokens.expand(image_embeds.shape[0], -1, -1)
-        query_tokens = query_tokens.to(self.qformer.dtype) # ours
-        image_embeds = image_embeds.to(self.qformer.dtype) # ours
         query_outputs = self.qformer(
             query_embeds=query_tokens,
             encoder_hidden_states=image_embeds,
@@ -497,7 +493,6 @@ class RBLNBlip2ForConditionalGeneration(RBLNModel, RBLNDecoderOnlyGenerationMixi
 
         if query_output.dtype != image_embeds.dtype:
             query_output = query_output.to(image_embeds.dtype)
-            query_output = query_output.to(self.dtype) # ours
 
         language_model_inputs = self.language_projection(query_output)
 
@@ -523,7 +518,6 @@ class RBLNBlip2ForConditionalGeneration(RBLNModel, RBLNDecoderOnlyGenerationMixi
         special_image_mask = special_image_mask.unsqueeze(-1).expand_as(inputs_embeds).to(inputs_embeds.device)
         language_model_inputs = language_model_inputs.to(inputs_embeds.device, inputs_embeds.dtype)
         inputs_embeds = inputs_embeds.masked_scatter(special_image_mask, language_model_inputs)
-        inputs_embeds = inputs_embeds.to(self.language_model.dtype) # ours
 
         inputs = {"inputs_embeds": inputs_embeds, "attention_mask": attention_mask}
         if not self.language_model.config.is_encoder_decoder:
